Move validation tests from Java 20 preview to Java 21
Record patterns and pattern matching for switch are finalized in Java 21,
so no need to test them as preview features in earlier versions.
