+++ org.jacoco.core.test.validation.java21/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.jacoco.core.test.validation.java20</name>
+	<name>org.jacoco.core.test.validation.java21</name>
 	<comment></comment>
 	<projects>
 	</projects>
